Revert "Added CSS elements, simplified element names"
