+++ b2nfrontend/src/index.css
@@ -1,5 +1,4 @@
-.body {
-  background-image: "./backgroundImg.gif";
+body {
   margin: 0;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
@@ -8,7 +7,7 @@
   -moz-osx-font-smoothing: grayscale;
 }
 
-.code {
+code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
